refactor: redirect new scenario route to playground
